Export ping logic to packet package
